x86: Extract 32-bit specific vspace functions in vspace_32
