+++ include/arch/x86/arch/fastpath/fastpath.h
@@ -30,15 +30,15 @@ switchToThread_fp(tcb_t *thread, pde_t *pd, pde_t stored_hw_asid)
      * for layout of gdt_data */
     /* update the GDT_TLS entry with the thread's TLS_BASE address */
     base = getRegister(thread, TLS_BASE);
-    gdt_entry_gdt_data_ptr_set_base_low(ia32KSgdt + GDT_TLS, base);
-    gdt_entry_gdt_data_ptr_set_base_mid(ia32KSgdt + GDT_TLS,  (base >> 16) & 0xFF);
-    gdt_entry_gdt_data_ptr_set_base_high(ia32KSgdt + GDT_TLS, (base >> 24) & 0xFF);
+    gdt_entry_gdt_data_ptr_set_base_low(x86KSgdt + GDT_TLS, base);
+    gdt_entry_gdt_data_ptr_set_base_mid(x86KSgdt + GDT_TLS,  (base >> 16) & 0xFF);
+    gdt_entry_gdt_data_ptr_set_base_high(x86KSgdt + GDT_TLS, (base >> 24) & 0xFF);
 
     /* update the GDT_IPCBUF entry with the thread's IPC buffer address */
     base = thread->tcbIPCBuffer;
-    gdt_entry_gdt_data_ptr_set_base_low(ia32KSgdt + GDT_IPCBUF, base);
-    gdt_entry_gdt_data_ptr_set_base_mid(ia32KSgdt + GDT_IPCBUF,  (base >> 16) & 0xFF);
-    gdt_entry_gdt_data_ptr_set_base_high(ia32KSgdt + GDT_IPCBUF, (base >> 24) & 0xFF);
+    gdt_entry_gdt_data_ptr_set_base_low(x86KSgdt + GDT_IPCBUF, base);
+    gdt_entry_gdt_data_ptr_set_base_mid(x86KSgdt + GDT_IPCBUF,  (base >> 16) & 0xFF);
+    gdt_entry_gdt_data_ptr_set_base_high(x86KSgdt + GDT_IPCBUF, (base >> 24) & 0xFF);
 
     ksCurThread = thread;
 }
@@ -115,7 +115,7 @@ fastpath_restore(word_t badge, word_t msgInfo, tcb_t *cur_thread)
         /* No-one (including us) is using the FPU, so we assume it
          * is currently disabled */
     }
-    tss_ptr_set_esp0(&ia32KStss, ((uint32_t)cur_thread) + 0x4c);
+    tss_ptr_set_esp0(&x86KStss, ((uint32_t)ksCurThread) + 0x4c);
     cur_thread->tcbArch.tcbContext.registers[EFLAGS] &= ~0x200;
     if (likely(hasDefaultSelectors(cur_thread))) {
         asm volatile("\
