+++ src/arch/x86/model/statedata.c
@@ -23,13 +23,13 @@ interrupt_t ia32KScurInterrupt VISIBLE;
 /* ==== proper read/write kernel state ==== */
 
 /* Task State Segment (TSS), contains currently running TCB in ESP0 */
-tss_t ia32KStss VISIBLE;
+tss_t x86KStss VISIBLE;
 
 /* Global Descriptor Table (GDT) */
-gdt_entry_t ia32KSgdt[GDT_ENTRIES];
+gdt_entry_t x86KSgdt[GDT_ENTRIES];
 
 /* The top level ASID table */
-asid_pool_t* ia32KSASIDTable[BIT(asidHighBits)];
+asid_pool_t* x86KSASIDTable[BIT(asidHighBits)];
 
 /*
  * Current thread whose state is installed in the FPU, or NULL if
@@ -40,10 +40,10 @@ tcb_t *ia32KSfpuOwner VISIBLE;
 /* ==== read-only kernel state (only written during bootstrapping) ==== */
 
 /* CPU Cache Line Size */
-uint32_t ia32KScacheLineSizeBits;
+uint32_t x86KScacheLineSizeBits;
 
 /* Interrupt Descriptor Table (IDT) */
-idt_entry_t ia32KSidt[IDT_ENTRIES];
+idt_entry_t x86KSidt[IDT_ENTRIES];
 
 /* A valid initial FPU state, copied to every new thread. */
 user_fpu_state_t ia32KSnullFpuState ALIGN(MIN_FPU_ALIGNMENT);
